Goal and football added.
Created a football prefab with a rigidbody from Jack R's Model.

Created a goal prefab with a script for detecting footballs from Jack R's Model.

Created a ground prefab that has a trigger that detects when the football passes through it and then adds to the miss counter or takes a life, whatever we decide to do with misses.